move pages around, get rid of inline styling, fix nav routing
