+++ worldkind/src/components/Image/index.js
@@ -1,4 +1,7 @@
 import styled from "styled-components"
 
-const Image = styled.img``
+const Image = styled.img`
+  width: ${props => props.width};
+  height: ${props => props.height};
+`
 export default Image
